Working CUDA kernels. Moved transpose kernels to a dedicated subdirectory.
