--- dali/kernels/common/transpose_gpu.h
@@ -1,36 +1,0 @@
-// Copyright (c) 2020, NVIDIA CORPORATION. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef DALI_KERNELS_COMMON_TRANSPOSE_GPU_H_
-#define DALI_KERNELS_COMMON_TRANSPOSE_GPU_H_
-
-#include "dali/core/tensor_view.h"
-
-namespace dali {
-namespace kernels {
-
-template <typename Out, typename In>
-class TransposeBatchGPU {
- public:
-  KernelRequirements Setup(const TensorListShape<> &in_shape, span<const int> permutation) {
-  }
-
- private:
-
-};
-
-}  // namespace kernels
-}  // namespace dali
-
-#endif  // DALI_KERNELS_COMMON_TRANSPOSE_GPU_H_
